pyproject.toml
[project]
name = "verify-mock-verifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "fastmcp>=2.5.1",
    "pillow>=11.2.1",
]
